getSubtreeUnrooted {ggtree}	R Documentation

getSubtreeUnrooted

Description

Get all subtrees of specified node. This includes all ancestors and relatives of node and return named list of subtrees.

Usage

getSubtreeUnrooted(tree, node)

Arguments

tree
ape phylo tree object

node
is the tree node id from which the subtrees are derived.

Value

named list of subtrees with the root id of subtree and list of node id's making up subtree.

[Package ggtree version 1.9.1 Index]