MRCA {ggtree}	R Documentation

MRCA

Description

Find Most Recent Common Ancestor among a vector of tips

Usage

MRCA(obj, tip)

Arguments

obj
supported tree object or ggplot object

tip
a vector of mode numeric or character specifying the tips

Value

MRCA of two or more tips

Author(s)

Guangchuang Yu

[Package ggtree version 1.9.1 Index]