%+>% {ggtree}	R Documentation

%+>%

Description

update data with tree info (y coordination and panel)

Usage

p %+>% data

Arguments

p
tree view

data
data.frame

Details

add tree information to an input data. This function will setup y coordination and panel info for data used in facet_plot and geom_faceet

Value

updated data.frame

Author(s)

Guangchuang Yu

References

G Yu, TTY Lam, H Zhu, Y Guan (2018). Two methods for mapping and visualizing associated data on phylogeny using ggtree. Molecular Biology and Evolution, 35(2):3041-3043. https://doi.org/10.1093/molbev/msy194

[Package ggtree version 3.2.1 Index]