annotation_image {ggtree}	R Documentation

annotation_image

Description

annotation taxa with images

Usage

annotation_image(tree_view, img_info, width = 0.1, align = TRUE,
  linetype = "dotted", linesize = 1, offset = 0)

Arguments

tree_view
tree view

img_info
data.frame with first column of taxa name and second column of image names

width
width of the image to be plotted in image

align
logical

linetype
line type if align = TRUE

linesize
line size if align = TRUE

offset
offset of image from the tree view

Value

tree view

Author(s)

Guangchuang Yu

[Package ggtree version 1.9.1 Index]