applyLayoutDaylight {ggtree}	R Documentation

applyLayoutDaylight

Description

Apply the daylight alorithm to adjust the spacing between the subtrees and tips of the specified node.

Usage

applyLayoutDaylight(df, node_id)

Arguments

df
tree data.frame

node_id
is id of the node from which daylight is measured to the other subtrees.

Value

list with tree data.frame with updated layout using daylight algorithm and max_change angle.

[Package ggtree version 1.9.1 Index]