facet_data {ggtree}	R Documentation

facet_data

Description

extract data used in facet_plot or geom_facet

Usage

facet_data(tree_view, panel)

Arguments

tree_view
ggtree object

panel
data plotted in specific panel. If only one dataset used in the panel, return the data frame, else return a list of data frames.

Value

data frame or a list of data frames

Author(s)

Guangchuang Yu

References

G Yu, TTY Lam, H Zhu, Y Guan (2018). Two methods for mapping and visualizing associated data on phylogeny using ggtree. Molecular Biology and Evolution, 35(2):3041-3043. https://doi.org/10.1093/molbev/msy194

[Package ggtree version 3.2.1 Index]