geom_balance {ggtree}	R Documentation

geom_balance

Description

highlights the two direct descendant clades of an internal node

Usage

geom_balance(node, fill = "steelblue", color = "white", alpha = 0.5,
  extend = 0, extendto = NULL)

Arguments

node
selected node (balance) to highlight

fill
color fill

color
color to outline highlights and divide balance

alpha
alpha (transparency)

extend
extend xmax of the rectangle

extendto
extend xmax to extendto

Details

Particularly useful when studying neighboring clades. Note that balances that correspond to multichotomies will not be displayed.

Value

ggplot2

Author(s)

Justin Silverman

References

J. Silverman, et al. A phylogenetic transform enhances analysis of compositional microbiota data. (in preparation)

[Package ggtree version 1.9.1 Index]