geom_label2 {ggtree}	R Documentation

geom_label2

Description

geom_label2 support aes(subset) via setup_data

Usage

geom_label2(mapping = NULL, data = NULL, ..., position = "identity",
  family = "sans", parse = FALSE, nudge_x = 0, nudge_y = 0,
  label.padding = unit(0.25, "lines"), label.r = unit(0.15, "lines"),
  label.size = 0.25, na.rm = TRUE, show.legend = NA, inherit.aes = TRUE)

Arguments

mapping
the aesthetic mapping

data
A layer specific dataset - only needed if you want to override the plot defaults.

...
other arguments passed on to 'layer'

position
The position adjustment to use for overlapping points on this layer

family
sans by default, can be any supported font

parse
if TRUE, the labels will be passd into expressions

nudge_x
horizontal adjustment

nudge_y
vertical adjustment

label.padding
Amount of padding around label.

label.r
Radius of rounded corners.

label.size
Size of label border, in mm

na.rm
logical

show.legend
logical

inherit.aes
logical

Value

label layer

Author(s)

Guangchuang Yu

See Also

geom_label

[Package ggtree version 1.9.1 Index]