geom_motif {ggtree}	R Documentation

geom_motif

Description

geom layer to draw aligned motif

Usage

geom_motif(mapping, data, on, label, align = "centre", ...)

Arguments

mapping
aes mapping

data
data

on
gene to center (i.e. set middle position of the on gene to 0)

label
specify a column to be used to label genes

align
where to place gene label, default is 'centre' and can be set to 'left' and 'right'

...
additional parameters

Value

geom layer

Author(s)

Guangchuang Yu

[Package ggtree version 3.2.1 Index]