geom_nodelab {ggtree}	R Documentation

geom_nodelab

Description

add node label layer

Usage

geom_nodelab(
  mapping = NULL,
  nudge_x = 0,
  nudge_y = 0,
  geom = "text",
  hjust = 0.5,
  node = "internal",
  ...
)

Arguments

mapping
aes mapping

nudge_x
horizontal adjustment to nudge label

nudge_y
vertical adjustment to nudge label

geom
one of 'text', "shadowtext", 'label', 'image' and 'phylopic'

hjust
horizontal alignment, one of 0, 0.5 or 1

node
a character indicating which node labels will be displayed, it should be one of 'internal', 'external' and 'all'. If it is set to 'internal' will display internal node labels, 'external' will display the tip labels, and 'all' will display internal node and tip labels.

...
additional parameters, see also the additional parameters of geom_tiplab().

Value

geom layer

Author(s)

Guangchuang Yu

See Also

geom_tiplab()

[Package ggtree version 3.2.1 Index]