geom_text2 {ggtree}	R Documentation

geom_text2

Description

geom_text2 support aes(subset) via setup_data

Usage

geom_text2(mapping = NULL, data = NULL, ..., position = "identity",
  family = "sans", parse = FALSE, na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, nudge_x = 0, nudge_y = 0, check_overlap = FALSE)

Arguments

mapping
the aesthetic mapping

data
A layer specific dataset - only needed if you want to override he plot defaults.

...
other arguments passed on to 'layer'

position
The position adjustment to use for overlapping points on this layer

family
sans by default, can be any supported font

parse
if TRUE, the labels will be passd into expressions

na.rm
logical

show.legend
logical

inherit.aes
logical

nudge_x
horizontal adjustment

nudge_y
vertical adjustment

check_overlap
if TRUE, text that overlaps previous text in the same layer will not be plotted

Value

text layer

Author(s)

Guangchuang Yu

See Also

geom_text

[Package ggtree version 1.9.1 Index]