geom_tippoint {ggtree}	R Documentation

geom_tippoint

Description

add tip point

Usage

geom_tippoint(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

Arguments

mapping
Set of aesthetic mapping created by aes(). If inherit.aes = TRUE, the mapping can be inherited from the plot mapping as specified in the call to ggplot().

data
The data to be displayed in this layer. If 'NULL' (the default), the data is inherited from the plot data as specified in the call to 'ggplot()',

position
Position adjustment.

na.rm
logical. If 'FALSE' (the default), missing values are removed with a warning. If 'TRUE', missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? 'NA', the default, includes if any aesthetics are mapped. 'FALSE' never includes, and 'TRUE' always includes.

inherit.aes
logical (default is 'TRUE'). If 'FALSE', overrides the default aesthetics, rather then combining with them.

...
addtional parameters that passed on to this layer. These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3.

Value

tip point layer

Author(s)

Guangchuang Yu

[Package ggtree version 3.2.1 Index]