getNodeEuclDistances {ggtree}	R Documentation

Get the distances from the node to all other nodes in data.frame (including itself if in df)

Description

Get the distances from the node to all other nodes in data.frame (including itself if in df)

Usage

getNodeEuclDistances(df, node)

[Package ggtree version 1.9.1 Index]