getSubtree {ggtree}	R Documentation

getSubtree

Description

Get all children of node from tree, including start_node.

Usage

getSubtree(tree, node)

Arguments

tree
ape phylo tree object

node
is the tree node id from which the tree is derived.

Value

list of all child node id's from starting node.

[Package ggtree version 1.9.1 Index]