getSubtreeUnrooted.df {ggtree}	R Documentation

getSubtreeUnrooted

Description

Get all subtrees of node, as well as remaining branches of parent (ie, rest of tree structure as subtree) return named list of subtrees with list name as starting node id.

Usage

getSubtreeUnrooted.df(df, node)

Arguments

df
tree data.frame

node
is the tree node id from which the subtrees are derived.

Value

named list of subtrees with the root id of subtree and list of node id's making up subtree.

[Package ggtree version 1.9.1 Index]