getTreeArcAngles {ggtree}	R Documentation

getTreeArcAngles

Description

Find the right (clockwise rotation, angle from +ve x-axis to furthest subtree nodes) and left (anti-clockwise angle from +ve x-axis to subtree) Returning arc angle in [0, 2] (0 to 360) domain.

Usage

getTreeArcAngles(df, origin_id, subtree)

Arguments

df
tree data.frame

origin_id
node id from which to calculate left and right hand angles of subtree.

subtree
named list of root id of subtree (node) and list of node ids for given subtree (subtree).

Value

named list with right and left angles in range [0,2] i.e 1 = 180 degrees, 1.5 = 270 degrees.

[Package ggtree version 1.9.1 Index]