get_taxa_name {ggtree}	R Documentation

get_taxa_name

Description

get taxa name of a selected node (or tree if node=NULL) sorted by their position in plotting

Usage

get_taxa_name(tree_view = NULL, node = NULL)

Arguments

tree_view
tree view (i.e. the ggtree object). If tree_view is NULL, the last ggplot will be used.

node
internal node number to specify a clade. If NULL, using the whole tree

Details

This function extract an ordered vector of the tips from selected clade or the whole tree based on the ggtree() plot.

Value

ordered taxa name vector

Author(s)

Guangchuang Yu

Examples

tree <- rtree(30)
p <- ggtree(tree)
get_taxa_name(p)

[Package ggtree version 3.2.1 Index]