ggtree {ggtree}	R Documentation

visualizing phylogenetic tree and heterogenous associated data based on grammar of graphics ggtree provides functions for visualizing phylogenetic tree and its associated data in R.

Description

visualizing phylogenetic tree and heterogenous associated data based on grammar of graphics ggtree provides functions for visualizing phylogenetic tree and its associated data in R.

drawing phylogenetic tree from phylo object

Usage

ggtree(tr, mapping = NULL, layout = "rectangular", open.angle = 0,
  mrsd = NULL, as.Date = FALSE, yscale = "none", yscale_mapping = NULL,
  ladderize = TRUE, right = FALSE, branch.length = "branch.length",
  ndigits = NULL, ...)

Arguments

tr
phylo object

mapping
aes mapping

layout
one of 'rectangular', 'slanted', 'fan', 'circular', 'radial', 'equal_angle' or 'daylight'

open.angle
open angle, only for 'fan' layout

mrsd
most recent sampling date

as.Date
logical whether using Date class in time tree

yscale
y scale

yscale_mapping
yscale mapping for category variable

ladderize
logical

right
logical

branch.length
variable for scaling branch, if 'none' draw cladogram

ndigits
number of digits to round numerical annotation variable

...
additional parameter

Value

tree

Author(s)

Yu Guangchuang

Examples

require(ape)
tr <- rtree(10)
ggtree(tr)

[Package ggtree version 1.9.1 Index]