gheatmap {ggtree}	R Documentation

gheatmap

Description

append a heatmap of a matrix to right side of phylogenetic tree

Usage

gheatmap(p, data, offset = 0, width = 1, low = "green", high = "red",
  color = "white", colnames = TRUE, colnames_position = "bottom",
  colnames_angle = 0, colnames_level = NULL, colnames_offset_x = 0,
  colnames_offset_y = 0, font.size = 4, hjust = 0.5)

Arguments

p
tree view

data
matrix or data.frame

offset
offset of heatmap to tree

width
total width of heatmap, compare to width of tree

low
color of lowest value

high
color of highest value

color
color of heatmap cell border

colnames
logical, add matrix colnames or not

colnames_position
one of 'bottom' or 'top'

colnames_angle
angle of column names

colnames_level
levels of colnames

colnames_offset_x
x offset for column names

colnames_offset_y
y offset for column names

font.size
font size of matrix colnames

hjust
hjust for column names (0: align left, 0.5: align center, 1: align righ)

Value

tree view

Author(s)

Guangchuang Yu

[Package ggtree version 1.9.1 Index]