layoutDaylight {ggtree}	R Documentation

Equal daylight layout method for unrooted trees.

Description

#' @title

Usage

layoutDaylight(tree, branch.length)

Arguments

tree
phylo object

branch.length
set to 'none' for edge length of 1. Otherwise the phylogenetic tree edge length is used.

Value

tree as data.frame with equal angle layout.

References

The following aglorithm aims to implement the vague description of the "Equal-daylight Algorithm" in "Inferring Phylogenies" pp 582-584 by Joseph Felsenstein.

“' Leafs are subtrees with no children Initialise tree using equal angle algorithm tree_df = equal_angle(tree)

nodes = get list of nodes in tree_df breadth-first nodes = remove tip nodes.

“'

[Package ggtree version 1.9.1 Index]