nodebar {ggtree}	R Documentation

nodebar

Description

generate a list of bar charts for results of ancestral state reconstruction

Usage

nodebar(data, cols, color, alpha = 1, position = "stack")

Arguments

data
a data.frame of stats with an additional column of node number

cols
column of stats

color
color of bar

alpha
alpha

position
position of bar, one of 'stack' and 'dodge'

Value

list of ggplot objects

Author(s)

Guangchuang Yu

[Package ggtree version 1.9.1 Index]