nodepie {ggtree}	R Documentation

nodepie

Description

generate a list of pie charts for results of ancestral stat reconstruction

Usage

nodepie(data, cols, color, alpha = 1)

Arguments

data
a data.frame of stats with an additional column of node number

cols
column of stats

color
color of bar

alpha
alpha

Value

list of ggplot objects

Author(s)

Guangchuang Yu

[Package ggtree version 1.9.1 Index]