rotateTreePoints.df {ggtree}	R Documentation

rotateTreePoints.data.frame

Description

Rotate the points in a tree data.frame around a pivot node by the angle specified.

Usage

rotateTreePoints.df(df, pivot_node, nodes, angle)

Arguments

df
tree data.frame

pivot_node
is the id of the pivot node.

nodes
list of node numbers that are to be rotated by angle around the pivot_node

angle
in range [0,2], ie degrees/180, radians/pi

Value

updated tree data.frame with points rotated by angle

[Package ggtree version 3.2.1 Index]